+++ constants/src/constants.rs
@@ -2,8 +2,8 @@ pub const WINDOW_TITLE: &str = "Walpurgis";
 
 pub const WINDOW_SCALE: f32 = 2.0;
 
-pub const TILE_WIDTH: u32 = 8;
-pub const TILE_HEIGHT: u32 = 8;
+pub const TILE_WIDTH: usize = 8;
+pub const TILE_HEIGHT: usize = 8;
 
-pub const GRID_WIDTH: u32 = 20;
-pub const GRID_HEIGHT: u32 = 18;
+pub const GRID_WIDTH: usize = 20;
+pub const GRID_HEIGHT: usize = 18;
